+++ sync-for-commerce/src/sdk/models/operations/updateconfigtextsyncflow.ts
@@ -3,7 +3,7 @@
  */
 
 import { SpeakeasyBase, SpeakeasyMetadata } from "../../../internal/utils";
-import * as shared from "../shared";
+import * as shared from "../../../sdk/models/shared";
 import { AxiosResponse } from "axios";
 
 export class UpdateConfigTextSyncFlowResponse extends SpeakeasyBase {
@@ -12,6 +12,12 @@ export class UpdateConfigTextSyncFlowResponse extends SpeakeasyBase {
      */
     @SpeakeasyMetadata()
     contentType: string;
+
+    /**
+     * The request made is not valid.
+     */
+    @SpeakeasyMetadata()
+    errorMessage?: shared.ErrorMessage;
 
     /**
      * Success
@@ -29,5 +35,5 @@ export class UpdateConfigTextSyncFlowResponse extends SpeakeasyBase {
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 }