+++ sync-for-commerce/src/sdk/models/shared/datatypefeature.ts
@@ -9,7 +9,7 @@ import { Expose, Type } from "class-transformer";
 /**
  * Available Data types
  */
-export enum DataTypeFeatureDataTypes {
+export enum DataTypes {
     AccountTransactions = "accountTransactions",
     BalanceSheet = "balanceSheet",
     BankAccounts = "bankAccounts",
@@ -25,6 +25,7 @@ export enum DataTypeFeatureDataTypes {
     DirectCosts = "directCosts",
     DirectIncomes = "directIncomes",
     Invoices = "invoices",
+    ItemReceipts = "itemReceipts",
     Items = "items",
     JournalEntries = "journalEntries",
     Journals = "journals",
@@ -63,7 +64,7 @@ export class DataTypeFeature extends SpeakeasyBase {
      */
     @SpeakeasyMetadata()
     @Expose({ name: "dataType" })
-    dataType?: DataTypeFeatureDataTypes;
+    dataType?: DataTypes;
 
     @SpeakeasyMetadata({ elemType: SupportedFeature })
     @Expose({ name: "supportedFeatures" })